Move cw-abc curves to its own package cw-curves
