+++ Cargo.toml
@@ -88,6 +88,7 @@ cw-ownable = "0.5"
 
 cw-abc = { path = "./contracts/external/cw-abc", version = "*" }
 cw-admin-factory = { path = "./contracts/external/cw-admin-factory", version = "2.4.1" }
+cw-curves = { path = "./packages/cw-curves", version = "2.4.1" }
 cw-denom = { path = "./packages/cw-denom", version = "2.4.1" }
 cw-hooks = { path = "./packages/cw-hooks", version = "2.4.1" }
 cw-paginate-storage = { path = "./packages/cw-paginate-storage", version = "2.4.1" }
