+++ contracts/external/cw-abc/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod abc;
 pub(crate) mod commands;
 pub mod contract;
-pub mod curves;
 mod error;
 pub mod msg;
 mod queries;
