+++ packages/cw-curves/README.md
@@ -1,0 +1,6 @@
+# CosmWasm Curves
+
+This package provides the curves to be used for
+[cw-abc](../../contracts/external/cw-abc).
+
+It provides a framework for defining various types of curves, such as constant, linear, and square root curves. The library ensures precision in token operations allowing for the easy implementation of custom curves in CosmWasm-based applications.
