+++ packages/cw-curves/src/curve.rs
@@ -1,0 +1,73 @@
+use cosmwasm_schema::cw_serde;
+use cosmwasm_std::{Decimal as StdDecimal, Uint128};
+use rust_decimal::prelude::ToPrimitive;
+use rust_decimal::Decimal;
+
+use crate::utils::decimal;
+
+/// This defines the curves we are using.
+///
+/// I am struggling on what type to use for the math. Tokens are often stored as Uint128,
+/// but they may have 6 or 9 digits. When using constant or linear functions, this doesn't matter
+/// much, but for non-linear functions a lot more. Also, supply and reserve most likely have different
+/// decimals... either we leave it for the callers to normalize and accept a `Decimal` input,
+/// or we pass in `Uint128` as well as the decimal places for supply and reserve.
+///
+/// After working the first route and realizing that `Decimal` is not all that great to work with
+/// when you want to do more complex math than add and multiply `Uint128`, I decided to go the second
+/// route. That made the signatures quite complex and my final idea was to pass in `supply_decimal`
+/// and `reserve_decimal` in the curve constructors.
+pub trait Curve {
+    /// Returns the spot price given the supply.
+    /// `f(x)` from the README
+    fn spot_price(&self, supply: Uint128) -> StdDecimal;
+
+    /// Returns the total price paid up to purchase supply tokens (integral)
+    /// `F(x)` from the README
+    fn reserve(&self, supply: Uint128) -> Uint128;
+
+    /// Inverse of reserve. Returns how many tokens would be issued
+    /// with a total paid amount of reserve.
+    /// `F^-1(x)` from the README
+    fn supply(&self, reserve: Uint128) -> Uint128;
+}
+
+/// DecimalPlaces should be passed into curve constructors
+#[cw_serde]
+pub struct DecimalPlaces {
+    /// Number of decimal places for the supply token (this is what was passed in cw20-base instantiate
+    pub supply: u32,
+    /// Number of decimal places for the reserve token (eg. 6 for uatom, 9 for nstep, 18 for wei)
+    pub reserve: u32,
+}
+
+impl DecimalPlaces {
+    pub fn new(supply: u8, reserve: u8) -> Self {
+        DecimalPlaces {
+            supply: supply as u32,
+            reserve: reserve as u32,
+        }
+    }
+
+    pub fn to_reserve(self, reserve: Decimal) -> Uint128 {
+        let factor = decimal(10u128.pow(self.reserve), 0);
+        let out = reserve * factor;
+        // TODO: execute overflow better? Result?
+        out.floor().to_u128().unwrap().into()
+    }
+
+    pub fn to_supply(self, supply: Decimal) -> Uint128 {
+        let factor = decimal(10u128.pow(self.supply), 0);
+        let out = supply * factor;
+        // TODO: execute overflow better? Result?
+        out.floor().to_u128().unwrap().into()
+    }
+
+    pub fn from_supply(&self, supply: Uint128) -> Decimal {
+        decimal(supply, self.supply)
+    }
+
+    pub fn from_reserve(&self, reserve: Uint128) -> Decimal {
+        decimal(reserve, self.reserve)
+    }
+}
