+++ packages/cw-curves/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod curve;
+pub mod curves;
+pub mod utils;
+
+#[cfg(test)]
+mod tests;
+
+pub use curve::{Curve, DecimalPlaces};
